Some corrections
