+++ src/main/java/ru/mail/polis/ads/SolveTemplate.java
@@ -19,15 +19,15 @@ public final class SolveTemplate {
         // Write me
     }
 
-    private static class FastScanner {
+    public static class FastScanner {
         private final BufferedReader reader;
         private StringTokenizer tokenizer;
 
-        FastScanner(final InputStream in) {
+        public FastScanner(final InputStream in) {
             reader = new BufferedReader(new InputStreamReader(in));
         }
 
-        String next() {
+        public String next() {
             while (tokenizer == null || !tokenizer.hasMoreTokens()) {
                 try {
                     tokenizer = new StringTokenizer(reader.readLine());
@@ -38,7 +38,7 @@ public final class SolveTemplate {
             return tokenizer.nextToken();
         }
 
-        int nextInt() {
+        public int nextInt() {
             return Integer.parseInt(next());
         }
     }
